feat: 비즈니스 로직과 UI 로직 분리 및 전역 프리텐다드 폰트 설정
