+++ src/features/map/components/WeatherButton.tsx
@@ -1,7 +1,7 @@
 import styles from "./WeatherOverlay.module.css";
 import { getCurrentSeason } from "../utils/weather";
-import { useBottomSheetStore } from "@/common/hooks/useBottomSheetStore";
 import { WeatherIcon } from "./WeatherIcon";
+import { useShelterButtonInteraction } from "../hooks/useShelterButtonInteraction";
 import type { TodayWeatherResponse } from "../schemas/weather.schema";
 
 export type WeatherButtonProps = {
@@ -14,22 +14,7 @@ export type WeatherButtonProps = {
 export function WeatherButton({ label, weather, disabled, onClick }: WeatherButtonProps) {
   const season = getCurrentSeason();
   const seasonClass = season === "summer" ? styles.buttonSummer : styles.buttonWinter;
-  const isOpen = useBottomSheetStore(state => state.isOpen);
-  const translateY = useBottomSheetStore(state => state.translateY);
-  const expandToTop = useBottomSheetStore(state => state.expandToTop);
-  const collapseToBottom = useBottomSheetStore(state => state.collapseToBottom);
-
-  const handleShelterClick = () => {
-    if (!isOpen) return;
-
-    const isAtTop = translateY < 50;
-
-    if (isAtTop && collapseToBottom) {
-      collapseToBottom();
-    } else if (!isAtTop && expandToTop) {
-      expandToTop();
-    }
-  };
+  const { handleShelterClick } = useShelterButtonInteraction();
 
   return (
     <div className={`${styles.button} ${seasonClass}`}>
